cargo/manifest: remember whether target names were inferred
Add a boolean state to `Target` that tells us whether the name of the
target was inferred by Cargo, or whether it was directly specified in
the Manifest.

This value will be required in the future, to allow changing the
inferred names of targets, but retaining enough information to keep
backwards compatibility.