ci/contrib: do not fail on missing gh-pages
The current contrib-deploy hook fails if there is no `gh-pages` branch.
Change the CI order to disregard the old `gh-pages` branch first.

The `contrib` deploy-hook always creates a fresh `gh-pages` commit and
pushes it out. However, currently it relies on the old `gh-pages` branch
to exist, since it does not ignore errors when pruning it. Fortunately,
the code always creates a new orphan branch, since it does not want to
keep history for deployments. Therefore, we can simply change the order
of the calls and always clear the branch first, then relying on `git
worktree --orphan` to create a fresh orphan branch.